+ Weight-class filtering for BAs + Bug: HGR PSRs wrong

+++ megamek/src/megamek/MegaMek.java
@@ -104,7 +104,7 @@ public class MegaMek {
     /**
      * This function returns the memory used in the heap (heap memory - free
      * memory).
-     * 
+     *
      * @return memory used in kB
      */
     public static String getMemoryUsed() {
@@ -117,7 +117,7 @@ public class MegaMek {
     /**
      * This function redirects the standard error and output streams to the
      * given File name.
-     * 
+     *
      * @param logFileName
      *            The file name to redirect to.
      */
@@ -142,7 +142,7 @@ public class MegaMek {
      * Starts a dedicated server with the arguments in args. See
      * {@link megamek.server.DedicatedServer#start(String[])} for more
      * information.
-     * 
+     *
      * @param args
      *            the arguments to the dedicated server.
      */
@@ -156,7 +156,7 @@ public class MegaMek {
     /**
      * Attempts to start the GUI with the given name. If the GUI is unknown the
      * program will exit.
-     * 
+     *
      * @param guiName
      *            The name of the GUI, usually AWT or swing
      * @param args
@@ -179,7 +179,7 @@ public class MegaMek {
 
     /**
      * Return the Interface to the GUI specified by the name in guiName.
-     * 
+     *
      * @param guiName
      *            the name of the GUI, will be passed on to
      *            {@link #getGUIClassName(String)}.
@@ -223,7 +223,7 @@ public class MegaMek {
     /**
      * This function appends 'agrs: []', to the buffer, with a space separated
      * list of args[] elements between the brackets.
-     * 
+     *
      * @param buffer
      *            the buffer to append the list to.
      * @param args
@@ -244,7 +244,7 @@ public class MegaMek {
 
     /**
      * Prints the message to stdout and then exits with errorcode 1.
-     * 
+     *
      * @param message
      *            the message to be displayed.
      */
@@ -255,7 +255,7 @@ public class MegaMek {
 
     /**
      * Prints the message and flushes the output stream.
-     * 
+     *
      * @param message
      */
     private static void displayMessage(String message) {
@@ -288,7 +288,7 @@ public class MegaMek {
 
     /**
      * Returns the version of Megamek
-     * 
+     *
      * @return the version of Megamek as a string.
      */
     public static String getVersion() {
@@ -322,7 +322,7 @@ public class MegaMek {
 
         /**
          * Returns <code>true</code> if the dedicated server option was found
-         * 
+         *
          * @return true iff this is a dedicated server.
          */
         public boolean dedicatedServer() {
@@ -332,7 +332,7 @@ public class MegaMek {
         /**
          * Returns the GUI Name option value or <code>null</code> if it wasn't
          * set
-         * 
+         *
          * @return GUI Name option value or <code>null</code> if it wasn't set
          */
         public String getGuiName() {
@@ -342,7 +342,7 @@ public class MegaMek {
         /**
          * Returns the log file name option value or <code>null</code> if it
          * wasn't set
-         * 
+         *
          * @return the log file name option value or <code>null</code> if it
          *         wasn't set
          */
@@ -352,7 +352,7 @@ public class MegaMek {
 
         /**
          * Returns the the <code>array</code> of the unprocessed arguments
-         * 
+         *
          * @return the the <code>array</code> of the unprocessed arguments
          */
         public String[] getRestArgs() {
@@ -601,7 +601,7 @@ public class MegaMek {
                         w.write(",");
                         w.write(Integer.toString(unit.getYear()));
                         w.write(",");
-                        w.write(Integer.toString(unit.getTons()));
+                        w.write(Float.toString(unit.getTons()));
                         w.write(",");
                         if (unit.isClan()) {
                             w.write("Clan,");
